Create a video from images and audio

Given a vector of paths to images (preferably .jpgs or .pngs) and a flat list of Waves of equal length this function will create an .mp4 video file where each image is shown with its corresponding audio. Take a look at the readWave function if you want to import your audio files into R. Please be sure that all images have the same dimensions.

ari_stitch(images, audio, output = tempfile(fileext = ".mp4"),
  verbose = FALSE, cleanup = TRUE, ffmpeg_opts = "",
  divisible_height = FALSE, audio_codec = get_audio_codec(),
  video_codec = get_video_codec(), audio_bitrate = "192k",
  video_bitrate = NULL)

Arguments

images
A vector of paths to images.

audio
A list of Waves from tuneR.

output
A path to the video file which will be created.

verbose
print diagnostic messages. If > 1, then more are printed

cleanup
If TRUE, interim files are deleted

ffmpeg_opts
additional options to send to ffmpeg. This is an advanced option, use at your own risk

divisible_height
Make height divisible by 2, which may be required if getting "height not divisible by 2" error.

audio_codec
The audio encoder for the splicing. If this fails, try copy.

video_codec
The video encoder for the splicing. If this fails, see ffmpeg -codecs

audio_bitrate
Bit rate for audio. Passed to -b:a.

video_bitrate
Bit rate for video. Passed to -b:v.

Value

A logical value, with the attribute outfile for the output file.

Details

This function uses FFmpeg which you should be sure is installed before using this function. If running Sys.which("ffmpeg") in your R console returns an empty string after installing FFmpeg then you should set the path to FFmpeg on you computer to an environmental variable using Sys.setenv(ffmpeg = "path/to/ffmpeg"). The environmental variable will always override the result of Sys.which("ffmpeg").

Examples

if (have_ffmpeg_exec()) {
result = ari_stitch(
ari_example(c("mab1.png", "mab2.png")),
list(tuneR::noise(), tuneR::noise()))
}